+++ unit_test.cpp
@@ -20,7 +20,10 @@ public:
         : value{std::make_shared<ResourceUsage>()}
     {};
 
-    void set(std::string const & revocable_allocated, std::string const & non_revocable_allocated) {
+    void set(
+        std::string const & revocable_allocated, std::string const & non_revocable_allocated,
+        std::string const & revocable_mem_used, std::string const & non_revocable_mem_used
+    ) {
         value->Clear();
         auto * revocable_executor = value->add_executors();
         auto revocable_resources = Resources::parse(revocable_allocated);
@@ -29,6 +32,8 @@ public:
             mutable_resource->CopyFrom(parsed_resource);
             mutable_resource->mutable_revocable();  // mark as revocable
         }
+        auto revocable_stats = revocable_executor->mutable_statistics();
+        revocable_stats->set_mem_total_bytes(Bytes::parse(revocable_mem_used).get().bytes());
 
         auto * non_revocable_executor = value->add_executors();
         auto non_revocable_resources = Resources::parse(non_revocable_allocated);
@@ -37,6 +42,8 @@ public:
             mutable_resource->CopyFrom(parsed_resource);
             mutable_resource->clear_revocable();  // mark as non-revocable
         }
+        auto non_revocable_stats = non_revocable_executor->mutable_statistics();
+        non_revocable_stats->set_mem_total_bytes(Bytes::parse(non_revocable_mem_used).get().bytes());
     }
     Future<ResourceUsage> operator()() const {
         return *value;
@@ -70,14 +77,16 @@ private:
 
 TEST(UsageMockTest, test_set) {
     UsageMock mock;
-    mock.set("cpus:3;mem:62", "cpus:1;mem:48");
+    mock.set("cpus:3;mem:62", "cpus:1;mem:48", "50MB", "32MB");
 
     ResourceUsage usage = mock().get();
     Resources allocatedRevocable;
     Resources allocatedNonRevocable;
+    uint64_t memUsed{0};
     for(auto & executor: usage.executors()) {
         allocatedRevocable += Resources(executor.allocated()).revocable();
         allocatedNonRevocable += Resources(executor.allocated()).nonRevocable();
+        memUsed += executor.statistics().mem_total_bytes();
     }
 
     EXPECT_EQ(2, usage.executors_size());
@@ -85,15 +94,18 @@ TEST(UsageMockTest, test_set) {
     EXPECT_EQ(Bytes::parse("62MB").get(), allocatedRevocable.mem().get());
     EXPECT_EQ(1, allocatedNonRevocable.cpus().get());
     EXPECT_EQ(Bytes::parse("48MB").get(), allocatedNonRevocable.mem().get());
+    EXPECT_EQ(Bytes::parse("82MB").get().bytes(), memUsed);
 
     // must also work for copy
     UsageMock copy = mock;
-    mock.set("cpus:5;mem:127", "");
+    mock.set("cpus:5;mem:127", "", "0B", "0B");
     usage = copy().get();
 
     Resources copiedRevocable;
+    memUsed = 0;
     for(auto & executor: usage.executors()) {
         copiedRevocable += Resources(executor.allocated()).revocable();
+        memUsed += executor.statistics().mem_total_bytes();
     }
     EXPECT_EQ(5, copiedRevocable.cpus().get());
     EXPECT_EQ(Bytes::parse("127MB").get(), copiedRevocable.mem().get());
@@ -112,7 +124,7 @@ TEST(LoadMockTest, test_set) {
 TEST(ThresholdResourceEstimatorTest, test_noop) {
     UsageMock usage;
     LoadMock load;
-    ThresholdResourceEstimator estimator{load, Resources::parse("").get(), None(), None(), None()};
+    ThresholdResourceEstimator estimator{load, Resources::parse("").get(), None(), None(), None(), None()};
     estimator.initialize(usage);
     auto available_resources = estimator.oversubscribable().get();
     EXPECT_TRUE(available_resources.empty());
@@ -121,7 +133,11 @@ TEST(ThresholdResourceEstimatorTest, test_noop) {
 TEST(ThresholdResourceEstimatorTest, test_underutilized) {
     UsageMock usage;  // no usage at all
     LoadMock load;
-    ThresholdResourceEstimator estimator{load, Resources::parse("cpus(*):2;mem(*):512").get(), None(), None(), None()};
+    ThresholdResourceEstimator estimator{
+        load,
+        Resources::parse("cpus(*):2;mem(*):512").get(),
+        None(), None(), None(), None()
+    };
     estimator.initialize(usage);
     auto available_resources = estimator.oversubscribable().get();
     EXPECT_FALSE(available_resources.empty());
@@ -129,13 +145,13 @@ TEST(ThresholdResourceEstimatorTest, test_underutilized) {
     EXPECT_EQ(Bytes::parse("512MB").get(), available_resources.revocable().mem().get());
 
     // Usage of revocable resources reduces offers
-    usage.set("cpus(*):1.5;mem(*):128", "");
+    usage.set("cpus(*):1.5;mem(*):128", "", "0B", "0B");
     available_resources = estimator.oversubscribable().get();
     EXPECT_EQ(0.5, available_resources.revocable().cpus().get());
     EXPECT_EQ(Bytes::parse("384MB").get(), available_resources.revocable().mem().get());
 
     // Usage of non-revocable resources is ignored
-    usage.set("", "cpus(*):1.5;mem(*):128");
+    usage.set("", "cpus(*):1.5;mem(*):128", "0B", "0B");
     available_resources = estimator.oversubscribable().get();
     EXPECT_EQ(2.0, available_resources.revocable().cpus().get());
     EXPECT_EQ(Bytes::parse("512MB").get(), available_resources.revocable().mem().get());
@@ -150,7 +166,8 @@ TEST(ThresholdResourceEstimatorTest, test_load_threshold_exceeded) {
         Resources::parse("cpus(*):2;mem(*):512").get(),
         4,
         3,
-        2
+        2,
+        None()
     };
     estimator.initialize(usage);
 
@@ -164,9 +181,15 @@ TEST(ThresholdResourceEstimatorTest, test_load_threshold_exceeded) {
     load.set(4.0, 2.9, 1.9);
     available_resources = estimator.oversubscribable().get();
     EXPECT_TRUE(available_resources.empty());
+    load.set(10.0, 2.9, 1.9);
+    available_resources = estimator.oversubscribable().get();
+    EXPECT_TRUE(available_resources.empty());
 
     // test 5 minute threshold exceeded
     load.set(3.9, 3.0, 1.9);
+    available_resources = estimator.oversubscribable().get();
+    EXPECT_TRUE(available_resources.empty());
+    load.set(3.9, 10.0, 1.9);
     available_resources = estimator.oversubscribable().get();
     EXPECT_TRUE(available_resources.empty());
 
@@ -174,6 +197,45 @@ TEST(ThresholdResourceEstimatorTest, test_load_threshold_exceeded) {
     load.set(3.9, 2.9, 2.0);
     available_resources = estimator.oversubscribable().get();
     EXPECT_TRUE(available_resources.empty());
+    load.set(3.9, 2.9, 10.0);
+    available_resources = estimator.oversubscribable().get();
+    EXPECT_TRUE(available_resources.empty());
+}
+
+TEST(ThresholdResourceEstimatorTest, test_mem_threshold_exceeded) {
+    UsageMock usage;
+    usage.set("cpus(*):1.5;mem(*):128", "cpus(*):1.5;mem(*):128", "128MB", "128MB");
+    LoadMock load;
+    load.set(3.9, 2.9, 1.9);
+    ThresholdResourceEstimator estimator{
+        load,
+        Resources::parse("cpus(*):2;mem(*):512").get(),
+        None(),
+        None(),
+        None(),
+        Bytes::parse("384MB").get()
+    };
+    estimator.initialize(usage);
+
+    // test _not_ exceeded
+    auto available_resources = estimator.oversubscribable().get();
+    EXPECT_FALSE(available_resources.empty());
+
+    // test threshold reached
+    usage.set("cpus(*):1.5;mem(*):128", "cpus(*):1.5;mem(*):128", "384MB", "0B");
+    available_resources = estimator.oversubscribable().get();
+    EXPECT_TRUE(available_resources.empty());
+
+    usage.set("cpus(*):1.5;mem(*):128", "cpus(*):1.5;mem(*):128", "0B", "384MB");
+    available_resources = estimator.oversubscribable().get();
+    EXPECT_TRUE(available_resources.empty());
+
+    // test threshold exceeded in non-revocable + revocable
+    usage.set("cpus(*):1.5;mem(*):128", "cpus(*):1.5;mem(*):128", "256MB", "256MB");
+    available_resources = estimator.oversubscribable().get();
+    EXPECT_TRUE(available_resources.empty());
+
+    // TODO test behaviour in case ResourceStatistics are not provided
 }
 
 }